Turn on bidir. typechecking hints for `mmatch`.
This is a first attempt at using Coq 8.11's `&` marker to improve the
typechecking for `mmatch`. It should now use the expected type given in
`in .. as .. return ..` to typecheck the branches instead of inferring a
type for the branches—which most certainly contains unavailable binders—
and then aborting.

Some places need new type annotations for matches that previously didn't
need them but I would rather have more of those than random typechecking
failures.

This also introduces a `branch_cons` wrapper around `mcons` so that we
can turn on bidirectional typechecking *just* for that wrapper, thus
avoiding potential issues in other places where `mcons` is used.