Req. parametrized constructors in [declare_mind].
This is necessary to eventually support returning the inductive type
from [declare_mind]. For this, we want to return the constructors
without any parameters applied to them. However, we cannot reparametrize
the constructors in pure Coq. That, however, is necessary to keep the
connection between the user input and the declared inductive type we
want to return.

Roughly speaking, the reason is that we cannot go from [Params -> list C]
to [list (Params -> C)].