Make `declare_mind` return the inductive type.
This changes the type signature of `declare_mind`. Since the
newly-defined type doesn't exist in the initial global environment, this
also means that `declare_mind` cannot possibly be executed by `mrun`.
This was already true before because tactics in general are not even
allowed to change the environment. But now that we return the type this
problem has become visible to us through assertion failures during
typechecking.