Be careful to not destruct tuple in `netuple_nth`.
It turns out that we will call this function on the tuple of
constructors return by `inspect_mind`. And while we know that those
tuples have a very specific shape, we do not actually know the value.
Thus, we need to be careful to not destruct them unless we want to be
yelled at by the typechecker.